agents_api/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
bcrypt==4.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
distro==1.9.0
dnspython==2.7.0
duckduckgo_search==6.3.3
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.115.4
fastapi-cli==0.0.5
feedparser==6.0.11
filelock==3.16.1
Flask==3.0.3
frozendict==2.4.6
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.1.1
h11==0.14.0
html5lib==1.1
httpcore==1.0.6
httptools==0.6.4
httpx==0.27.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.6.1
joblib==1.4.2
lxml==5.3.0
lxml_html_clean==0.3.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
multitasking==0.0.11
newspaper4k==0.9.3.1
nltk==3.9.1
numpy==2.1.2
openai==1.52.2
pandas==2.2.3
passlib==1.7.4
peewee==3.17.7
phidata==2.5.25
pillow==11.0.0
platformdirs==4.3.6
primp==0.6.5
psycopg2-binary==2.9.10
pyasn1==0.6.1
pydantic==2.9.2
pydantic-settings==2.6.0
pydantic_core==2.23.4
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.16
pytz==2024.2
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
requests-file==2.1.0
rich==13.9.3
rsa==4.9
setuptools==74.1.2
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.36
starlette==0.41.2
tldextract==5.1.2
tomli==2.0.2
tqdm==4.66.6
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.0
watchfiles==0.24.0
webencodings==0.5.1
websockets==13.1
Werkzeug==3.0.4
yfinance==0.2.48
